backend/requirements.txt
SQLAlchemy
fastapi
pydantic
psycopg2
pycparser
python-dotenv
pydantic-settings
pydantic[dotenv]
alembic
uvicorn
